+++ src/components/ResumeBook.js
@@ -5,7 +5,7 @@ import { Register } from "./auth/Register";
 import { NavBar } from "./nav/NavBar";
 import { UserProvider } from "./users/UserProvider";
 import { ApplicationViews } from "./ApplicationViews";
-import { ExpProvider } from "./posts/ExpProvider";
+import { ExpProvider } from "./posts/exps/ExpProvider";
 import { SkillProvider } from "./posts/skills/SkillProvider";
 import { ProjectProvider } from "./posts/ProjectProvider";
